+++ desktop/modules/system-dashboard/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.1] - 2026-05-18
+
+### Fixed
+- 🐛 **同步阻塞**：`getDiskUsage()` 从 `execSync` 改为 `exec` + Promise 异步执行，不再阻塞主进程
+
 ## [1.0.0] - 2026-05-18
 
 ### Added